Switch to setuptools
This required a bit more of a lift than I might have wanted, but the result should be more maintainable in the long run.

In particular, I did a lot of rewrite to resource.py which may have broken it. It works for loading the images at the moment, but it might not work for text files.