Refactor nmcli logic
This changes the nmcli logic to use "nmcli device set <interface>
unmanaged" instead of modifying the config files.

This change also ignores a specific case where nmcli is unable to
contact NetworkManager, instead assuming that NM is not managing the
interface. This is useful if you're in a Distrobox that can't talk to
the outside NM.